Add experimental parser for alphabetical list of series
Constructing BeautifulSoup objects is time-consuming on resource-limited
devices, such as Raspberry Pi. Using the default getting list of all
series takes up to 50 seconds on Raspberry Pi3.

This patch adds an experimental parser that breaks easily on website change
but OTOH it's able to do the same job in less than 9 seconds. The option
to use this parser (disabled by default) is also added to addon settings.